Added precision for cannibalization demands
Floating point math wasn't as accurate as BigInt for calculating
the cannibalization demand quantity and rounding up.  For the base
case supply, using floats would result in rounding up to the next int
when integer math would say it should the the previous int from
next int.

This caused test data changes which is causing one test to fail.