Tests now pass with rounding fix
although we are filtering out all of the demand records that
changed with the fix for now.  Supporting code was added to that
we can write a new test to check a couple things on the records
that were changed to make sure they changed as we expected.

Before this fix, we were only testing the ac supply records, so
the supply test still passes since it is only the rc records that
are affected.  Really, we should filter the supply for records
that don't have rc forward or the rounding fix.  For the rounding
fixes, just make sure that the Tags contain a / for now to verify
that we switched to rationals. For the rc forwards, make sure that
we have :align-units.