Added updated data for current TAA
Working on getting usage examples setup
for complete TAA run on new pipeline.

Need to tie in taapost as a dependency
to allow the full gamut of pipeline
to be accessed from the usage example.

Future efforts will provide worked
examples as part of the documentation.

We need to tease out the inline comments
into a form that's amenable to docs.
Might be able to do that with marginalia,
dunno